Add a headless litehtml command-line tool
headless is a command-line tool that uses litehtml to render an HTML
document.  headless can only render HTML documents from the local file
system (i.e., it cannot load HTML documents over HTTP or HTTPS).  The
rendered document is saved to a PNG file in the current directory.

headless is meant to provide a way to quickly and easily render HTML
while working on the litehtml engine itself.

headless is not built by default as it requires cairo and fontconfig
for its container implementation (which is based on container_linux
with the GTK components removed).  Pass BUILD_HEADLESS=1 to CMake
to build headless.