Add parser performance tests
Add (well, fix) performance tests for the CSSParser and DocumentParser
classes. The performance tests are written using the Google Benchmark
framework, and use non-trivial CSS and HTML files to measure the
performance of the parsers.

The performance tests are disabled by default, and cam be enabled by
passing -DBUILD_PERFTEST=1 to CMake.